source/requirements.txt
flask>=2.0.0
werkzeug>=2.0.0
tensorflow>=2.6.0
pillow>=8.0.0
numpy>=1.19.0